Make headers install correctly with scikit-build-core, rename package
