+++ components/core/tests/cpp_generation_test.cc
@@ -8,7 +8,7 @@
 #include "test_expressions.h"  //  Symbolic test functions.
 
 #define MATH_SPAN_EIGEN_SUPPORT
-#include "wf_runtime/span_eigen.h"
+#include "wrenfold/span_eigen.h"
 
 // Declare custom numeric types before importing the generated code:
 namespace wf::numeric {
