+++ components/python/wrenfold/code_generation.py
@@ -182,7 +182,7 @@ CPP_PREAMBLE_TEMPLATE = \
 #include <cmath>
 #include <cstdint>
 
-#include <wf_runtime/span.h>
+#include <wrenfold/span.h>
 
 namespace {namespace} {{
