+++ examples/bspline/bspline_test.cc
@@ -6,7 +6,7 @@
 #include "wf_test_support/test_macros.h"
 
 #define MATH_SPAN_EIGEN_SUPPORT
-#include "wf_runtime/span_eigen.h"
+#include "wrenfold/span_eigen.h"
 
 #include "bspline_numerical.h"
